extract spec_scanner
